requirements.txt
python-telegram-bot==20.7
psycopg[binary]==3.1.13
pymysql==1.1.0
reportlab==4.0.7
python-dotenv==1.0.0
schedule==1.2.0
